# Conflicts:
#	ifs/mods/topic/Dictionary.sh

+++ ifs/mods/add/dicts.sh
@@ -2,16 +2,23 @@
 # -*- ENCODING: UTF-8 -*-
 
 [ -z "$DM" ] && source /usr/share/idiomind/default/c.conf
+
 function dicts() {
+    dlg=0
     cmsg() {
-        sleep 5
-        if [ ! -e "$DC_s/topics_first_run" ]; then
-            source "$DS/ifs/mods/cmns.sh"
-            msg_2 "$(gettext "You may need to configure the list of Internet resources. \nDo you want to do this now?")" \
-            info "$(gettext "Yes")" "$(gettext "Cancel")" "$(gettext "Information")"
-            if [ $? = 0 ]; then "$DS_a/Dics/cnfg.sh" 6; fi
-            echo $lgtl > "$DC_a/dict/.dict"
+        if [ ! -e "$DT/dicts" ]; then
+            touch "$DT/dicts"
+            sleep 5
+            if [ ! -e "$DC_s/topics_first_run" ]; then
+                source "$DS/ifs/mods/cmns.sh"
+                msg_2 "$(gettext "You may need to configure the list of Internet resources. \nDo you want to do this now?")" \
+                info "$(gettext "Yes")" "$(gettext "Cancel")" "$(gettext "Information")"
+                if [ $? = 0 ]; then "$DS_a/Dics/cnfg.sh" 6; fi
+                echo $lgtl > "$DC_a/dict/.dict"
+            fi
+            rm -f "$DT/dicts"
         fi
+        return 0
     }
     s=0
     if [ ! -d "$DC_d" -o ! -d "$DC_a/dict/disables" ]; then
@@ -24,8 +31,10 @@ function dicts() {
     if  [ ! -f "$DC_a/dict/.dict" ]; then s=1
         echo -e "$lgtl" > "$DC_a/dict/.dict"
     fi
-    if ! ls "$DC_d"/* 1> /dev/null 2>&1; then cmsg; fi
-    if  [[ `sed -n 1p "$DC_a/dict/.dict"` != $lgtl ]] ; then cmsg; fi
+    if ! ls "$DC_d"/* 1> /dev/null 2>&1; then dlg=1; fi
+    if  [[ `sed -n 1p "$DC_a/dict/.dict"` != $lgtl ]] ; then dlg=1; fi
+    
+    [[ ${dlg} = 1 ]] && cmsg
 }
 
 dicts &
